r.stream.[segment|stats]: use double for easting/northing
git-svn-id: https://svn.osgeo.org/grass/grass-addons@69568 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ grass7/raster/r.stream.segment/stream_vector.c
@@ -5,7 +5,7 @@ int create_sector_vector(char *out_vector, int number_of_streams, int radians)
     int i, j, k;
     int r, c, d;
     int start, stop;
-    float northing, easting;
+    double northing, easting;
     struct Map_info Out;
     struct line_pnts *Segments;
     struct line_cats *Cats;
@@ -221,7 +221,7 @@ int create_segment_vector(char *out_vector, int number_of_streams,
 {
     int i, k;
     int r, c, d;
-    float northing, easting;
+    double northing, easting;
     struct Map_info Out;
     struct line_pnts *Segments;
     struct line_cats *Cats;
